// services/api/src/business-modelling/enums/bm-activity-type.enum.ts
export enum BMActivityType {
  BUSINESS_IDEA = 'BUSINESS_IDEA',
  PROBLEM_SOLUTION = 'PROBLEM_SOLUTION',
  CUSTOMER_ANALYSIS = 'CUSTOMER_ANALYSIS',
  VALUE_PROPOSITION = 'VALUE_PROPOSITION',
  MARKET_ANALYSIS = 'MARKET_ANALYSIS',
  COMPETITOR_ANALYSIS = 'COMPETITOR_ANALYSIS',
  BUSINESS_MODEL = 'BUSINESS_MODEL',
  PRODUCT_SERVICE = 'PRODUCT_SERVICE',
  PRICING = 'PRICING',
  REVENUE = 'REVENUE',
  COST_STRUCTURE = 'COST_STRUCTURE',
  CAPITAL = 'CAPITAL',
  OPERATIONS = 'OPERATIONS',
  MARKETING_SALES = 'MARKETING_SALES',
  FINANCIAL_MODEL = 'FINANCIAL_MODEL',
  CASH_FLOW = 'CASH_FLOW',
  BREAK_EVEN = 'BREAK_EVEN',
  ROI = 'ROI',
  RISK_ANALYSIS = 'RISK_ANALYSIS',
  GROWTH = 'GROWTH',
  BUSINESS_PLAN = 'BUSINESS_PLAN',
  PITCH = 'PITCH',
  CASE_STUDY = 'CASE_STUDY',
  PRACTICAL_MODEL = 'PRACTICAL_MODEL',
}